refactor: major refactor of the codebase
