+++ scripts/evaluate_avg.py
@@ -9,7 +9,7 @@ from tqdm import tqdm
 import matplotlib.pyplot as plt
 
 from utils.evaluate import get_sample_image
-from utils.data import get_loaders
+from schemas.components.dataset import get_loaders
 import torch
 from utils.metrics import display_metrics, display_grouped_metrics
